separated components from code
